pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "test-squadron-discord-bot"
version = "1.0.0"
description = "A Discord bot for managing TEST Squadron community functions"
authors = [{name = "ChromeNinja", email = "t.riley.garrett@gmail.com"}]
license = {text = "Proprietary"}
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.1,<3.0.0",
    "PyYAML>=6.0,<7.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "aiohttp>=3.8.4,<4.0.0",
    "beautifulsoup4>=4.12.2,<5.0.0",
    "lxml>=4.9.3,<5.0.0",
    "aiosqlite>=0.20.0,<1.0.0",
    "aiolimiter>=1.1.0,<2.0.0",
    "PyNaCl==1.5.0",
    "psutil>=5.9.0,<6.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "requests>=2.28.0,<3.0.0",
    "sqlalchemy[asyncio]>=2.0.0,<3.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
    "types-PyYAML",
    "types-beautifulsoup4",
    "pytest-cov>=4.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
show-fixes = true
src = ["."]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "YTT",    # flake8-2020
    "S",      # flake8-bandit
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T20",    # flake8-print
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PD",     # pandas-vet
    "PGH",    # pygrep-hooks
    "PL",     # pylint
    "TRY",    # tryceratops
    "NPY",    # numpy
    "RUF",    # ruff-specific
]

ignore = [
    "S101",   # assert used
    "T20",    # print found
    "COM812", # missing-trailing-comma (conflicts with formatter)
    "ISC001", # single-line-implicit-string-concatenation (conflicts with formatter)
    "E501",   # line-too-long - handled by formatter, not linter
    "PLC0415", # import-outside-top-level - deferred imports are intentional for performance
    "PLR2004", # magic-value-comparison - too noisy for constants
    "TRY003", # raise-vanilla-args - too restrictive
    "TRY401", # verbose-log-message - exc_info is often intentional
    "S608",   # hardcoded-sql-expression - SQL queries are intentional
    "ARG001", # unused-function-argument - often needed for interface compatibility
    "ARG002", # unused-method-argument - often needed for interface compatibility
    "TRY300", # try-consider-else - stylistic preference
    "TRY301", # raise-within-try - sometimes intentional for control flow
    "PLR0911", # too-many-return-statements - some functions need many returns
    "S311",   # suspicious-non-cryptographic-random-usage - random delays don't need crypto
    "RUF006", # asyncio-dangling-task - fire-and-forget tasks are intentional
    "N806",   # non-lowercase-variable-in-function - constants in functions are fine
    # DTZ003 (call-datetime-utcnow) is NOT ignored - use datetime.now(timezone.utc) instead
    "B904",   # raise-without-from-inside-except - not always needed
    "S110",   # try-except-pass - sometimes intentional for optional operations
    "SIM105", # suppressible-exception - contextlib.suppress isn't always clearer
    "PLW0603", # global-statement - intentional for singletons
    "SIM102", # collapsible-if - nested ifs can be clearer
    "SIM117", # multiple-with-statements - separate with statements can be clearer
    "RUF001", # ambiguous-unicode-character-string - emoji/special chars intentional
    "RUF002", # ambiguous-unicode-character-docstring - special chars in docs fine
    "RUF003", # ambiguous-unicode-character-comment - special chars in comments fine
    "PTH123", # builtin-open - Path.open() not always necessary
    "PGH003", # blanket-type-ignore - specific codes not always known
    "TRY002", # raise-vanilla-class - Exception is fine for simple cases
    "SIM103", # needless-bool - explicit returns can be clearer
    "PLW0602", # global-variable-not-assigned - intentional for module-level state
    "PLW2901", # redefined-loop-name - intentional reassignment in loops
    "B027",   # empty-method-without-abstract-decorator - intentional base implementations
    "E402",   # module-import-not-at-top-of-file - path manipulation needed first
    "S603",   # subprocess-without-shell-equals-true - shell=False is intentional
    "S105",   # hardcoded-password-string - false positive on URL variable names
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "TID252", "ARG", "PLR0913", "PLR0915", "F811", "RUF012", "B007", "SIM116", "ERA001"]
"web/backend/tests/**/*" = ["S101", "TID252", "ARG", "ERA001"]
"web/backend/tests/conftest.py" = ["PTH108"]
"web/backend/routes/**/*" = ["B008", "PLR0913", "PLR0912", "PLR0915"]  # FastAPI Depends pattern + complexity
"web/backend/core/**/*" = ["B008", "PLR0913", "F811"]  # FastAPI Depends pattern
"cogs/**/*" = ["PLR0913", "PLR0912", "PLR0915", "ARG003"]  # Discord cogs often have complex command handlers
"helpers/**/*" = ["PLR0913", "PLR0912", "PLR0915"]  # Helper functions may need many args
"services/**/*" = ["PLR0913", "PLR0912", "PLR0915", "ERA001"]  # Service methods may be complex
"bot.py" = ["PLR0912", "PLR0915"]  # Main bot file is complex
"scripts/**/*" = ["PLR0912", "PLR0915"]  # Scripts may be complex
"tools/**/*" = ["PLR0912", "PLR0915"]  # Tools may be complex
"utils/**/*" = ["PLR0913"]  # Utils may need many args
"verification/**/*" = ["PLR0912", "PLR0915"]  # Verification is complex
"**/migration*.py" = []  # Migration files are fine

[tool.pytest.ini_options]
testpaths = ["tests", "web/backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=bot",
    "--cov=cogs",
    "--cov=helpers",
    "--cov=services",
    "--cov=utils",
    "--cov=verification",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=44",
    "-m", "not integration",
    "-q",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["bot", "cogs", "helpers", "services", "utils", "verification"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "discord.*",
    "aiohttp.*",
    "bs4.*",
    "lxml.*",
    "aiosqlite.*",
    "aiolimiter.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
